- modules/m_gline.c: time_t warning (moved to %lu with a typecast).
git-svn-id: http://svn.ratbox.org/svnroot/trunk@7617 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/m_gline.c
@@ -683,13 +683,13 @@ log_gline(struct Client *source_p,
       return;
     }
 
-  ircsprintf(buffer, "\"%s\",\"%s\",\"%s %s\",\"%s\",%ld\n",
+  ircsprintf(buffer, "\"%s\",\"%s\",\"%s %s\",\"%s\",%lu\n",
 	     user,
 	     host,
 	     reason,
 	     timebuffer,
 	     oper_nick,
-	     CurrentTime);
+	     (unsigned long) CurrentTime);
 
   if (fbputs(buffer,out) == -1)
     {
